% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_marking_sheet.R
\name{build_marking_sheet}
\alias{build_marking_sheet}
\title{Build marking sheet from essays downloaded from Blackboard}
\usage{
build_marking_sheet(dir_list)
}
\arguments{
\item{dir_list}{A list (or vector) of paths to essays}
}
\value{
A data.frame of essay IDs
}
\description{
Build marking sheet from essays downloaded from Blackboard
}
