% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_emails.R
\name{extract_emails}
\alias{extract_emails}
\title{Extract email addresses from Blackboard user output}
\usage{
extract_emails(path_blackboard_email_file)
}
\arguments{
\item{path_blackboard_email_file}{Path to Blackboard user information file (in CSV format)}
}
\value{
A string containing the email addresses, separated by ';'
}
\description{
Extract email addresses from Blackboard user output
}
